refactor(reply): package delivery and media modules
