+++ music_reply.py
@@ -24,7 +24,7 @@ from latentsync_reply import (
 )
 from runtime.media.media_paths import configured_media_path
 from runtime.media.music_duration import MUSIC_DURATION_OPTIONS, normalize_music_duration as _normalize_music_duration
-from reply_media import (
+from runtime.reply.reply_media import (
     ReplyMediaError,
     assemble_complete_video_delivery,
     render_reply_video,
