+++ tts/delivery.py
@@ -15,7 +15,7 @@ from functools import lru_cache
 from pathlib import Path
 
 from latentsync_reply import LatentSyncReplyError, resolve_ffmpeg_executable
-from reply_delivery import ReplyDeliveryPlan
+from runtime.reply.reply_delivery import ReplyDeliveryPlan
 from voice_direction import VoiceDirectionError, validate_short_instruction
 
 from .contracts import TTSConfig
